requirements.txt
asgiref==3.5.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
idna==3.6
inflection==0.5.1
numpy==1.26.4
pygini==1.0.1
python-dotenv==0.21.0
pytz==2022.6
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.3
tzdata==2022.6
urllib3==2.2.1
whitenoise==6.6.0
